% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomass_models}
\alias{biomass_models}
\title{Biomass models}
\format{
A \code{tibble}
}
\usage{
biomass_models
}
\description{
Biomass models available in \code{silviculture} package. If you would like
to suggest new models, please open a new issue.
}
\keyword{datasets}
