% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{calc_accumulated_trees}
\alias{calc_accumulated_trees}
\title{Calculates number of trees until reaching a maximum number of trees}
\usage{
calc_accumulated_trees(ntrees, cumtrees, max_trees)
}
\arguments{
\item{ntrees}{Number of trees per hectare}

\item{cumtrees}{Accumulated trees and sorted from thickest to thinner diameter}

\item{max_trees}{Number to trees to calculate dominant metric}
}
\value{
A numeric vector
}
\description{
Calculates number of trees until reaching a maximum number of trees
}
\keyword{internal}
