% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{calc_dominant_height}
\alias{calc_dominant_height}
\title{Calculates dominant height}
\usage{
calc_dominant_height(nmax, ntress, height)
}
\arguments{
\item{nmax}{Index of first diametric class with > 100 trees; if there are no
100 trees, it is the index of the maximum}

\item{ntress}{Number of trees per hectare}

\item{height}{Height of the diametric class}
}
\value{
A numeric vector
}
\description{
Calculates dominant height
}
\keyword{internal}
