% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{calc_dominant_metric}
\alias{calc_dominant_metric}
\title{Calculates dominant height or dominant diameter}
\usage{
calc_dominant_metric(nmax, ntress, metric, max_trees = 100)
}
\arguments{
\item{nmax}{Index of first diametric class with > 100 trees; if there are no
100 trees, it is the index of the maximum}

\item{ntress}{Number of trees per hectare}

\item{metric}{Height of the diametric class or diameter}

\item{max_trees}{Number to trees to calculate dominant metric}
}
\value{
A numeric vector
}
\description{
Calculates dominant height or dominant diameter
}
\keyword{internal}
