% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{calc_n_optimal}
\alias{calc_n_optimal}
\title{Calculates number of plots for optimal allocation with constant cost}
\usage{
calc_n_optimal(students_t, pj, sj, max_n, max_error)
}
\arguments{
\item{students_t}{Student's T value}

\item{pj}{proportion of each stratum}

\item{sj}{variance of each stratum}

\item{max_n}{maximum number of plots in the area}

\item{max_error}{maximum allowed absolute error}
}
\value{
A length-one numeric vector
}
\description{
Calculates number of plots for optimal allocation with constant cost
}
\keyword{internal}
