% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{calc_n_simple}
\alias{calc_n_simple}
\title{Calculates number of plots using Student's T}
\usage{
calc_n_simple(students_t, max_n, cv, max_error)
}
\arguments{
\item{students_t}{Student's T value}

\item{max_n}{maximum number of plots in the area}

\item{cv}{coefficient of variation}

\item{max_error}{maximum allowed relative error}
}
\value{
A length-one numeric vector
}
\description{
Calculates number of plots using Student's T
}
\keyword{internal}
