% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-height.R
\name{eq_hd_aitor2025}
\alias{eq_hd_aitor2025}
\title{Estimates tree height from DBH}
\usage{
eq_hd_aitor2025(
  bioregion = "mediterranean",
  origin = "natural",
  mixture = "pure"
)
}
\arguments{
\item{bioregion}{The biogreopgrahic region of the species. Available options are:
\code{mediterranean}, \code{atlantic}, \code{alpine}, and \code{macaronesian}}

\item{origin}{The origin of the stand. Available options are: \code{natural} and
\code{plantation}}

\item{mixture}{The mixture of the species. Available options are: \code{pure} and
\code{mix}}
}
\value{
A numeric vector with predicted height
}
\description{
This function is made to be used in \link{silv_predict_height}. It implements the h-d equations
developed in Aitor et al. (2025). These equations had been adjusted using the Spanish
National Forest Inventory, and therefore, they should be used only within Spain. A wide
range of tree species are available (90).
}
\details{
Details...#TODO
}
\examples{
1 + 1 #TODO
}
\references{
Reference #TODO
}
