% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry-summary.R
\name{Inventory}
\alias{Inventory}
\title{Inventory class}
\usage{
Inventory(dclass_metrics = data.frame(), group_metrics = data.frame())
}
\arguments{
\item{dclass_metrics}{A data.frame summarised by diametric class with variables
such as plot_id, species, dclass, height, ntrees, ntrees_ha, h0, dg, and g_ha.}

\item{group_metrics}{A data.frame summarised by plot and species with variables
such as plot_id, species, d_mean, d_median, d_sd, dg, h_mean, h_median, h_sd,
h_lorey, h0, ntrees, ntrees_ha, g_ha, and spacing.}
}
\value{
An S7 \code{Inventory} object, which contains the inventory data.
}
\description{
A list containing forest inventory data summaries. It includes data by
diametric class (\code{dclass_metrics}) and by plot and species (\code{group_metrics}).
}
