% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventory_samples}
\alias{inventory_samples}
\title{Forest inventory samples}
\format{
A `tibble`
}
\usage{
inventory_samples
}
\description{
Inventory data from Spanish National Forest Inventory
}
\keyword{datasets}
