% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{plot}
\alias{plot}
\title{Plot an object}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{Object to plot.}

\item{...}{Other arguments passed to methods.}
}
\value{
Usually called for side-effects (producing a plot).
}
\description{
Generic for plotting objects.
}
