% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{SampleSize}
\alias{SampleSize}
\title{SampleSize class}
\usage{
SampleSize(sampling_res = list(), sampling_opts = list())
}
\arguments{
\item{sampling_res}{A list containing the results of sample size calculations.}

\item{sampling_opts}{A list containing the input options of sample size calculations.}
}
\value{
An S7 \code{SampleSize} object
}
\description{
A list containing the summary data of \link{silv_sample_size}. It includes input
data (sampling_opts) and output data (sampling_res)
}
