% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdi_coefficients}
\alias{sdi_coefficients}
\title{SDI beta coefficients}
\format{
A \code{tibble}
\describe{
\item{article_id}{Character. Short identifier of the source article.}
\item{title}{Character. Full title of the source article.}
\item{doi_url}{Character. DOI URL of the source article.}
\item{country}{Character. Country where the study was conducted.}
\item{region}{Character. Region within the country.}
\item{species}{Character. Scientific name of the tree species.}
\item{beta}{Numeric. Beta coefficient for SDI calculation.}
}
}
\usage{
sdi_coefficients
}
\description{
Specific beta coefficients for Reineke's Stand Density Index (SDI) per
species, country, and region.
}
\keyword{datasets}
