% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdimax_models}
\alias{sdimax_models}
\title{Maximum stand density index (SDImax) models}
\format{
A \code{tibble} with 88 rows and 13 variables:
\describe{
\item{article_id}{Character. Identifier of the article.}
\item{title}{Character. Title of the article.}
\item{doi_url}{Character. DOI URL of the article.}
\item{country}{Character. Country where the study was conducted.}
\item{species}{Character. Tree species scientific name.}
\item{model_name}{Character. Name of the model/equation variant (e.g. "basic", "P1", "MXT3").}
\item{a0}{Numeric. Coeffient a0.}
\item{a1}{Numeric. Coeffient a1 (0 if not used/applicable).}
\item{b0}{Numeric. Coeffient b0.}
\item{b1}{Numeric. Coeffient b1 (0 if not used/applicable).}
\item{aic}{Numeric. Akaike Information Criterion.}
\item{pseudo_r2}{Numeric. Pseudo R-squared value.}
\item{q_index}{Numeric. Q index value.}
}
}
\usage{
sdimax_models
}
\description{
Coefficients for calculating maximum stand density index (SDImax) from
Rodríguez de Prado (2020).
}
\references{
Rodríguez-de-Prado, M., et al. (2020). Potential climatic influence on maximum stand carrying capacity for 15 Mediterranean coniferous and broadleaf species. Forest Ecology and Management, 458, 117824.
}
\keyword{datasets}
