% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-stand-density.R
\name{silv_density_sdi_class}
\alias{silv_density_sdi_class}
\title{Classifies the Stand Density Index}
\usage{
silv_density_sdi_class(sdi, max_sdi, classify = TRUE)
}
\arguments{
\item{sdi}{A numeric vector representing the Stand Density Index.}

\item{max_sdi}{A numeric vector representing the maximum SDI for the species/site.}

\item{classify}{A logical value indicating whether to classify the values into density classes
(default is \code{TRUE}). If \code{FALSE}, it returns the relative SDI as a percentage.}
}
\value{
A character vector with the density classes if \code{classify = TRUE}, or a numeric vector
with the relative SDI percentage if \code{classify = FALSE}.
}
\description{
Classifies the Stand Density Index (SDI) into density classes or calculates the relative SDI
percentage based on USDA thresholds.
}
\details{
The option \code{classify = TRUE} will use the \code{max_sdi} value to classify the SDI into
four competitive and growth conditions: low density (<24\%), moderate density (24-35\%),
high density (34-55\%), and extremely high density (>55\%).
}
\examples{
## calculate SDI for a Pinus sylvestris stand (max 990)
sdi_val <- silv_density_sdi(ntrees = 800, dg = 23.4)

## check base classification
silv_density_sdi_class(sdi = sdi_val, max_sdi = 990)

## get relative SDI percentage
silv_density_sdi_class(sdi = sdi_val, max_sdi = 990, classify = FALSE)
}
\references{
USDA Forest Service. (n.d.). Stand Density Index.
https://www.fs.usda.gov/Internet/FSE_DOCUMENTS/stelprdb5270993.pdf
}
