% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-stand-level.R
\name{silv_site_factor}
\alias{silv_site_factor}
\title{Calculates the Site Factor (SF)}
\usage{
silv_site_factor(species, d0, h0)
}
\arguments{
\item{species}{Character vector. Scientific names of the tree species.}

\item{d0}{Numeric vector. Dominant diameter.}

\item{h0}{Numeric vector. Dominant height.}
}
\value{
A numeric vector representing the Site Factor. \code{NA} for unsupported species.
}
\description{
The site factor is a variable used in growth equations based on the
dominant height and dominant dbh.
}
\examples{
silv_site_factor(species = "Pinus sylvestris", d0 = 25, h0 = 15)
}
\references{
Aguirre, A., et al. (2022).
}
