% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory-summary.R
\name{silv_summary_mortality}
\alias{silv_summary_mortality}
\title{Summarize plot data by mortality status}
\usage{
silv_summary_mortality(
  data,
  plot_id,
  dead,
  expan = NULL,
  g = NULL,
  diameter = NULL
)
}
\arguments{
\item{data}{A data frame or tibble of tree-level data}

\item{plot_id}{Unquoted column name with the plot identifier}

\item{dead}{Unquoted column name indicating if the tree is dead (e.g., TRUE/FALSE or 1/0)}

\item{expan}{Unquoted column name with the expansion factor (trees/ha). If \code{NULL}, it is assumed each row represents 1 tree/ha.}

\item{g}{Unquoted column name with the basal area per tree (m²/ha). If \code{NULL}, \code{diameter} must be provided.}

\item{diameter}{Unquoted column name with the diameter (cm). Used to calculate basal area if \code{g} is \code{NULL}.}
}
\value{
A tibble with plot-level mortality summaries.
}
\description{
Calculates the number of trees and basal area per hectare for live and dead trees in a plot.
}
\examples{
library(dplyr)
inventory_samples |>
  mutate(expan = silv_density_ntrees_ha(1, 10)) |>
  mutate(is_dead = sample(c(TRUE, FALSE), n(), replace = TRUE, prob = c(0.1, 0.9))) |>
  silv_summary_mortality(plot_id, is_dead, expan, diameter = diameter)
}
