% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-tree-level.R
\name{silv_tree_bas}
\alias{silv_tree_bas}
\title{Calculate Basal Area Smaller (BAS)}
\usage{
silv_tree_bas(
  data,
  plot_id,
  tree_id,
  diameter = NULL,
  expansion_factor = NULL,
  basal_area_ha = NULL,
  units = "cm"
)
}
\arguments{
\item{data}{A data frame containing tree measurements.}

\item{plot_id}{Unquoted column name with the plot identifier.}

\item{tree_id}{Unquoted column name with the tree identifier.}

\item{diameter}{Unquoted column name with the DBH (optional).}

\item{expansion_factor}{Unquoted column name with the expansion factor (optional).}

\item{basal_area_ha}{Unquoted column name with the basal area per hectare. If provided, \code{diameter} and \code{expansion_factor} are ignored.}

\item{units}{The units of the diameter.}
}
\value{
A numeric vector with the BAS for each tree.
}
\description{
Calculates the Basal Area Smaller (BAS) for each tree in each plot.
}
