% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-tree-level.R
\name{silv_tree_basal_area_ha}
\alias{silv_tree_basal_area_ha}
\title{Calculate Tree Basal Area per Hectare}
\usage{
silv_tree_basal_area_ha(
  diameter = NULL,
  basal_area = NULL,
  expansion_factor,
  units = "cm"
)
}
\arguments{
\item{diameter}{A numeric vector of tree DBH. Optional if \code{basal_area} is provided.}

\item{basal_area}{A numeric vector of tree basal area in m2. Optional if \code{diameter} is provided.}

\item{expansion_factor}{A numeric vector representing the tree expansion factor.}

\item{units}{The units of the diameter (one of \code{mm}, \code{cm}, \code{dm}, or \code{m}).}
}
\value{
A numeric vector with the basal area per hectare (m2/ha).
}
\description{
Calculates the tree basal area expanded to hectare.
}
\examples{
silv_tree_basal_area_ha(diameter = 20, expansion_factor = 127)
}
