% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-tree-level.R
\name{silv_tree_circumference}
\alias{silv_tree_circumference}
\title{Calculate Tree Circumference}
\usage{
silv_tree_circumference(diameter, units = "cm")
}
\arguments{
\item{diameter}{Numeric vector of diameters or diameter classes}

\item{units}{The units of the diameter (one of \code{mm}, \code{cm}, \code{dm}, or \code{m}).}
}
\value{
A numeric vector with the circumference in the same units.
}
\description{
Calculate Tree Circumference
}
\examples{
silv_tree_circumference(20)
}
