% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-tree-level.R
\name{silv_tree_coordinates}
\alias{silv_tree_coordinates}
\title{Calculate Relative and Absolute Tree Coordinates}
\usage{
silv_tree_coordinates(
  distance,
  bearing,
  x_center = NULL,
  y_center = NULL,
  dist_units = "m",
  bearing_units = "grad"
)
}
\arguments{
\item{distance}{Distance from plot center.}

\item{bearing}{Bearing from plot center.}

\item{x_center}{Plot center X coordinate (optional).}

\item{y_center}{Plot center Y coordinate (optional).}

\item{dist_units}{Units of the distance (one of \code{mm}, \code{cm}, \code{dm}, \code{m}).}

\item{bearing_units}{Units of the bearing (\code{degree}, \code{grad}, \code{radian}).}
}
\value{
A data frame with relative coordinates (\code{x_rel}, \code{y_rel}) and, if plot
centers are provided, absolute coordinates (\code{x_abs}, \code{y_abs}).
}
\description{
Calculate Relative and Absolute Tree Coordinates
}
\examples{
silv_tree_coordinates(distance = 10, bearing = 100)
}
