% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-tree-level.R
\name{silv_tree_crown_ratio}
\alias{silv_tree_crown_ratio}
\title{Calculate Tree Crown Ratio}
\usage{
silv_tree_crown_ratio(height, height_base_crown)
}
\arguments{
\item{height}{Total tree height.}

\item{height_base_crown}{Height to the base of the live crown.}
}
\value{
A numeric vector with the crown ratio.
}
\description{
Calculate Tree Crown Ratio
}
\examples{
silv_tree_crown_ratio(height = 20, height_base_crown = 8)
}
