% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-tree-level.R
\name{silv_tree_mean_dbh}
\alias{silv_tree_mean_dbh}
\title{Calculate Mean DBH or derive it from Perimeter}
\usage{
silv_tree_mean_dbh(dbh1 = NULL, dbh2 = NULL, perimeter = NULL, units = "cm")
}
\arguments{
\item{dbh1}{A numeric vector with the first DBH measurement.}

\item{dbh2}{A numeric vector with the second DBH measurement.}

\item{perimeter}{A numeric vector with the perimeter measurement.}

\item{units}{The units of the inputs (one of \code{mm}, \code{cm}, \code{dm}, or \code{m}).}
}
\value{
A numeric vector with the mean DBH.
}
\description{
Calculates the mean diameter at breast height (DBH) from two perpendicular
measurements or derives it from the perimeter.
}
\examples{
silv_tree_mean_dbh(dbh1 = 20, dbh2 = 22)
silv_tree_mean_dbh(perimeter = 65)
}
