% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-tree-level.R
\name{silv_tree_slenderness}
\alias{silv_tree_slenderness}
\title{Calculate Tree Slenderness}
\usage{
silv_tree_slenderness(diameter, height, d_units = "cm", h_units = "m")
}
\arguments{
\item{diameter}{A numeric vector of tree DBH.}

\item{height}{A numeric vector of tree height.}

\item{d_units}{Units of the diameter.}

\item{h_units}{Units of the height.}
}
\value{
A numeric vector with the tree slenderness.
}
\description{
Calculate Tree Slenderness
}
\examples{
silv_tree_slenderness(diameter = 20, height = 15)
}
