% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silviculture-package.R
\docType{package}
\name{silviculture-package}
\alias{silviculture}
\alias{silviculture-package}
\title{silviculture: Utility Functions for Forest Inventory and Silviculture}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Perform common dendrometry operations such as inventory preparing, and inventory data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cidree.github.io/silviculture/}
}

}
\author{
\strong{Maintainer}: Adrián Cidre González \email{adrian.cidre@gmail.com}

Authors:
\itemize{
  \item Aitor Vázquez Veloso \email{aitor.vazquez.veloso@uva.es}
}

}
\keyword{internal}
