% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{plot_SimpleSampleSize}
\alias{plot_SimpleSampleSize}
\title{Plot Sample Size vs Error}
\arguments{
\item{x}{An object of class \code{SimpleSampleSize} containing sampling options and results.}

\item{min_error}{A numeric value specifying the minimum relative error to consider (default is 0.01).}

\item{max_error}{A numeric value specifying the maximum relative error to consider (default is 0.5).}
}
\value{
A \code{ggplot} object representing the relationship between error and sample size.
}
\description{
This method generates a plot showing how the required sample size varies with the maximum allowed relative error.
}
