% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_factor_models}
\alias{site_factor_models}
\title{Site Factor models}
\format{
A \code{tibble} with 24 rows and 7 variables:
\describe{
\item{species}{Character. Scientific name of the tree species.}
\item{species_code}{Numeric. Species numeric code.}
\item{d_ref}{Numeric. Reference diameter (Dref).}
\item{model}{Character. Model shape and expanded parameter.}
\item{param_a}{Numeric. Parameter a.}
\item{param_b}{Numeric. Parameter b.}
\item{param_c}{Numeric. Parameter c.}
}
}
\usage{
site_factor_models
}
\description{
Coefficients for calculating the Site Factor (SF) based on Aguirre et al. (2022).
The site factor is a variable that can be used in growth equations and is based on
the dominant height and dominant dbh.
}
\references{
Aguirre, A. et al. (2022).
}
\keyword{datasets}
