% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{weighted_sd}
\alias{weighted_sd}
\title{Calculates weighted standard deviation}
\usage{
weighted_sd(var, wt)
}
\arguments{
\item{var}{An object containing the values whose weighted median is to be computed}

\item{wt}{A numerical vector of weights the same length as x giving the weights to use for elements of x}
}
\value{
A length-one numeric vector
}
\description{
Calculates weighted standard deviation
}
\keyword{internal}
