% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-stand-density.R
\name{silv_density_sdi}
\alias{silv_density_sdi}
\title{Calculates the Stand Density Index}
\usage{
silv_density_sdi(ntrees, dg, beta = 1.605)
}
\arguments{
\item{ntrees}{Numeric vector with number of trees of the diameter class per
hectare. If \code{ntrees = NULL}, the function will assume that each diameter
corresponds to only one tree}

\item{dg}{Numeric vector of quadratic mean diameters}

\item{beta}{The Stand Density Index exponent (default is \code{1.605}).}
}
\value{
A numeric vector representing the absolute SDI.
}
\description{
The Stand Density Index (SDI) is the relationship between the average tree size and
density of trees per hectare.
}
\details{
The SDI has different interpretations depending on the species, location, and also
the management type (even-aged, uneven-aged...). The value of maximum SDI must
be determined from the literature and used carefully. The \code{beta} exponent allows
adjustments for different species or mixed stands.
}
\examples{
## calculate SDI for a Pinus sylvestris stand (beta = 1.605)
silv_density_sdi(ntrees = 800, dg = 23.4)

## calculate SDI with custom beta
silv_density_sdi(ntrees = 800, dg = 23.4, beta = 1.7)
}
\references{
Reineke, L. H. (1933). Perfecting a stand-density index for even-aged forests.
Journal of Agricultural Research, 46(7), 627-638.
}
