% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{silv_diametric_class}
\alias{silv_diametric_class}
\title{Classify diameters in classes}
\usage{
silv_diametric_class(
  x,
  dmin = 7.5,
  dmax = NULL,
  class_length = 5,
  include_lowest = TRUE,
  return_intervals = FALSE
)
}
\arguments{
\item{x}{A numeric vector of diameters}

\item{dmin}{The minimum inventory diameter in centimeters}

\item{dmax}{The maximum inventory diameter in centimeters. Values that
are greater than `dmax` are included in the greatest class}

\item{class_length}{The length of the class in centimeters}

\item{include_lowest}{Logical. If TRUE (the default), the intervals are
`[dim1, dim2)`. If FALSE, the intervals are `(dim1, dim2]`}

\item{return_intervals}{If FALSE, it returns the intervals. Otherwise (the
default), it returns the class center}
}
\value{
A numeric vector
}
\description{
Classifies the measured diameters into classes of a specified length
}
\examples{
library(dplyr)
inventory_samples |>
  mutate(dclass = silv_diametric_class(diameter))
}
