% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-height.R
\name{silv_predict_height}
\alias{silv_predict_height}
\title{Estimates tree height (m) from DBH (cm)}
\usage{
silv_predict_height(diameter, model, quiet = FALSE)
}
\arguments{
\item{diameter}{Numeric vector with diameters in cm (DBH).}

\item{model}{A ModelHD object. An object configured via the \code{eq_hd_*()} family of functions.
Currently only \code{\link[=eq_hd_vazquez_veloso_2025]{eq_hd_vazquez_veloso_2025()}} is available.}

\item{quiet}{A logical value. If TRUE, suppresses any informational messages.}
}
\value{
A numeric vector with predicted heights (m).
}
\description{
Estimates total tree height (m) using height-diameter (h-d) equations. Currently, only models developed
for Spain are available.
}
\details{
The function estimates total tree height (in meters) using diameter at breast height (in centimeters),
and may require additional information depending on the specific model. See each model’s documentation for details.
}
\examples{
# 1. Load the S7 ModelHD object for Pinus pinaster
model <- eq_hd_vazquez_veloso_2025("Pinus pinaster")

# 2. Vector-based calculation: predict tree heights from diameters
diameters <- c(20, 25, 30)
predicted_heights <- silv_predict_height(diameter = diameters, model = model)
print(predicted_heights)

# 3. Dataset-based tutorial: apply to a forest inventory data frame
inventory <- data.frame(
  tree_id = 1:3,
  species = c("Pinus pinaster", "Pinus pinaster", "Pinus pinaster"),
  dbh_cm  = c(18.5, 22.1, 29.4)
)

# Apply prediction and append a new column to the dataset
inventory$height_m <- silv_predict_height(
  diameter = inventory$dbh_cm,
  model    = model
)
print(inventory)
}
\references{
References for the models available:
\itemize{
\item \strong{\code{\link[=eq_hd_vazquez_veloso_2025]{eq_hd_vazquez_veloso_2025()}}}: Vázquez-Veloso, A., Yang, S.-I., Bullock, B.P., Bravo, F., 2025. One model to rule them all:
A nationwide height–diameter model for 91 Spanish forest species. Forest Ecology and Management 595, 122981.
https://doi.org/10.1016/j.foreco.2025.122981
}
}
\seealso{
\code{\link[=eq_hd_vazquez_veloso_2025]{eq_hd_vazquez_veloso_2025()}}
}
