% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-height.R
\name{silv_predict_height}
\alias{silv_predict_height}
\title{Estimates tree height from DBH}
\usage{
silv_predict_height(diameter, model, quiet = FALSE)
}
\arguments{
\item{diameter}{Numeric vector with diameters in cm}

\item{model}{A function. A function with the structure \code{eq_hd_*()} with
additional arguments depending on the model. Currently only \code{\link[=eq_hd_aitor2025]{eq_hd_aitor2025()}}
is available.}

\item{quiet}{A logical value. If TRUE, suppresses any informational messages.}
}
\value{
A numeric vector with predicted height
}
\description{
This function estimates tree height using h-d equations. Currently, only equations
for Spain are available.
}
\details{
Details...#TODO
}
\examples{
1 + 1 #TODO
}
\references{
Reference #TODO
}
\seealso{
\code{\link[=eq_hd_aitor2025]{eq_hd_aitor2025()}}
}
