% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{silv_sample_size}
\alias{silv_sample_size}
\title{Calculate sample size}
\usage{
silv_sample_size(
  x,
  plot_size = 100,
  total_area = 15,
  max_error = 0.05,
  conf_level = 0.95,
  max_iter = 1000,
  quiet = FALSE
)
}
\arguments{
\item{x}{vector of field survey}

\item{plot_size}{a numeric vector of length one with plot size in squared meters}

\item{total_area}{total area of the study area in squared meters}

\item{max_error}{maximum allowed error}

\item{conf_level}{confidence level}

\item{max_iter}{maximum number of iteration to find the plot size}

\item{quiet}{if TRUE, messages will be supressed}
}
\value{
SampleSize object
}
\description{
Calculate sample size
}
\examples{
1 + 1 # TODO
}
