% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory-summary.R
\name{silv_summary_species}
\alias{silv_summary_species}
\title{Summarize plot data by species}
\usage{
silv_summary_species(
  data,
  plot_id,
  species,
  expan = NULL,
  g = NULL,
  diameter = NULL,
  top_n = 3
)
}
\arguments{
\item{data}{A data frame or tibble of tree-level data}

\item{plot_id}{Unquoted column name with the plot identifier}

\item{species}{Unquoted column name with the species identifier}

\item{expan}{Unquoted column name with the expansion factor (trees/ha). If \code{NULL}, it is assumed each row represents 1 tree/ha.}

\item{g}{Unquoted column name with the basal area per tree (m²/ha). If \code{NULL}, \code{diameter} must be provided.}

\item{diameter}{Unquoted column name with the diameter (cm). Used to calculate basal area if \code{g} is \code{NULL}.}

\item{top_n}{Number of top species to pivot into columns (default: 3). If \code{0}, no pivoting is performed.}
}
\value{
A tibble with species-level summaries.
}
\description{
Calculates the number of trees and basal area per hectare for each species
in a plot, and optionally provides the top species by basal area.
}
\examples{
library(dplyr)
inventory_samples |>
  mutate(expan = silv_density_ntrees_ha(1, 10)) |>
  silv_summary_species(plot_id, species, expan, diameter = diameter, top_n = 3)
}
